requirements.txt
pytorch
torchvision
numpy
pillow
setuptools
pytest
wandb
